+++ core/src/com/mygdx/game/GameObjekts/SpaceShipParts/SpaceShipCocpit.java
@@ -2,7 +2,7 @@ package com.mygdx.game.GameObjekts.SpaceShipParts;
 
 public class SpaceShipCocpit extends ShipModule{
 
-    public SpaceShipCocpit(ModuleType moduleType, String name, double capacity, double fill, double cost, double baseFailureRate){
+    public SpaceShipCocpit(ModuleType moduleType, String name, double capacity, double fill, double cost, int baseFailureRate){
         super(moduleType, name, capacity, cost);
